+++ R/fia_assign_strata.R
@@ -45,17 +45,24 @@ fia_assign_strata <- function(data_annualized, db) {
   chosen_evals <- pop_info |>
     dplyr::filter(EXPVOL & EXPCURR) |>
     dplyr::arrange(INVYR, START_INVYR, END_INVYR) |>
-    dplyr::filter(EVALID_YEAR >= 1999) |> # TODO: possibly not necessary because of handling of specific states below
     dplyr::select(-INVYR) |>
-    fia_split_composite_ids()
-
-  # Handle NM and WY. They list early FHM inventories, but they don't work, so
-  # dropping. (modified from rFIA
-  # https://github.com/doserjef/rFIA/blob/ac9c8cb7c524935afeb25ef859ab422a2bb68044/R/getDesignInfo.R#L55C3-L62C4)
-  if (any(c(35, 56) %in% unique(chosen_evals$STATECD))) {
-    chosen_evals <- chosen_evals |>
-      dplyr::filter(!(.data$STATECD %in% c(35, 56) & .data$END_INVYR < 2001))
-  }
+    fia_split_composite_ids() |>
+    # Only keep EVALIDs after the start of the annual inventory in each state as
+    # an additional layer of precaution against matching "bad" EVALIDs.  This
+    # takes the place of some of the state-specific fixes implemented in rFIA
+    # here:
+    # https://github.com/doserjef/rFIA/blob/ac9c8cb7c524935afeb25ef859ab422a2bb68044/R/getDesignInfo.R#L55C3-L62C4)
+    dplyr::left_join(annual_inventory_start, by = "STATECD") |>
+    dplyr::select(
+      plot_ID,
+      EVALID,
+      EVALID_YEAR,
+      START_INVYR,
+      END_INVYR,
+      annual_inventory_start,
+      everything()
+    ) |>
+    dplyr::filter(START_INVYR >= annual_inventory_start)
 
   # Handle Texas.  Remove plots part of any EVALID associated with West/East
   # Texas
@@ -110,7 +117,7 @@ fia_assign_strata <- function(data_annualized, db) {
       EXPCURR = dplyr::if_else(is.na(EVALID), FALSE, EXPCURR),
       EXPVOL = dplyr::if_else(is.na(EXPVOL), FALSE, EXPVOL)
     ) |>
-    # remove cols that could be joined in from eval info
+    # remove cols that could easily be joined in from eval info later
     dplyr::select(
       -all_of(c(
         "UNITCD",